+++ example/lib/pages/chip_page.dart
@@ -1,5 +1,5 @@
-import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 import 'package:flutter/material.dart';
+import 'package:sbb_design_system_mobile/sbb_design_system_mobile.dart';
 
 import '../native_app.dart';
 
@@ -23,7 +23,7 @@ class _ChipPageState extends State<ChipPage> {
         const ThemeModeSegmentedButton(),
         const SizedBox(height: sbbDefaultSpacing),
         const SBBListHeader('Default'),
-        SBBGroup(
+        SBBContentBox(
           margin: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
           padding: const EdgeInsets.all(sbbDefaultSpacing),
           child: Column(
@@ -63,7 +63,7 @@ class _ChipPageState extends State<ChipPage> {
           ),
         ),
         const SBBListHeader('Disabled'),
-        SBBGroup(
+        SBBContentBox(
           margin: const EdgeInsets.symmetric(horizontal: sbbDefaultSpacing),
           padding: const EdgeInsets.all(sbbDefaultSpacing),
           child: Column(